split-table: split Table interface by behavior
