+++ repository/src/main/java/tech/ydb/yoj/repository/db/AbstractDelegatingTable.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import tech.ydb.yoj.databind.expression.FilterExpression;
 import tech.ydb.yoj.databind.expression.OrderExpression;
 import tech.ydb.yoj.repository.BaseDb;
+import tech.ydb.yoj.repository.db.cache.FirstLevelCache;
 import tech.ydb.yoj.repository.db.readtable.ReadTableParams;
 import tech.ydb.yoj.repository.db.statement.Changeset;
 
@@ -30,6 +31,11 @@ public abstract class AbstractDelegatingTable<T extends Entity<T>> implements Ta
     private Class<T> resolveEntityType() {
         return (Class<T>) (new TypeToken<T>(getClass()) {
         }).getRawType();
+    }
+
+    @Override
+    public FirstLevelCache getFirstLevelCache() {
+        return target.getFirstLevelCache();
     }
 
     @Override
